+++ lib/getch/filesystem.rb
@@ -4,3 +4,4 @@ module Getch
 end
 
 require_relative 'filesystem/ext4'
+require_relative 'filesystem/lvm'