+++ lib/getch/log.rb
@@ -3,6 +3,8 @@ require 'logger'
 module Getch
   class Log
     def initialize
+      @log_file = '/tmp/log_install.txt'
+      check_file
       init_log
       init_log_text
     end
@@ -29,6 +31,10 @@ module Getch
 
     private
 
+    def check_file
+      puts "Creating log at #{@log_file}" if ! File.exist? @log_file
+    end
+
     def init_log
       @logger = Logger.new(STDOUT)
       @logger.level = DEFAULT_OPTIONS[:verbose] ? Logger::DEBUG : Logger::INFO
@@ -38,7 +44,7 @@ module Getch
     end
 
     def init_log_text
-      @logger_text = Logger.new("/tmp/log_install.txt", 1)
+      @logger_text = Logger.new(@log_file, 1)
       @logger_text.level = Logger::DEBUG
       @logger_text.formatter = proc { |severity, datetime, progname, msg|
         "#{severity}, #{datetime}, #{msg}\n"